Revert "migration from kotlinx.datetime.Instant -> kotlin.time.Instant #1350"
This reverts commit 390b84daf53b125bee6a71c2cc5ca738a6307bb9.

+++ dataframe-arrow/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/ArrowWriterImpl.kt
@@ -237,7 +237,7 @@ internal class ArrowWriterImpl(
             is DateDayVector ->
                 column.convertToLocalDate()
                     .forEachIndexed { i, value ->
-                        value?.also { vector.set(i, value.toEpochDays().toInt()) }
+                        value?.also { vector.set(i, value.toEpochDays()) }
                             ?: vector.setNull(i)
                     }
